+++ p3/gen/export_wasi_http_handler/response.go
@@ -2,11 +2,12 @@ package export_wasi_http_handler
 
 import (
 	"errors"
-	"fmt"
+	"io"
 	"net/http"
 	"sync"
 
 	httpTypes "github.com/jamesstocktonj1/componentize-sdk/p3/gen/wasi_http_types"
+	internalhttp "github.com/jamesstocktonj1/componentize-sdk/p3/internal/wasihttp"
 	witTypes "go.bytecodealliance.org/pkg/wit/types"
 )
 
@@ -19,9 +20,8 @@ func newHttpResponseWriter() *responseHandler {
 }
 
 type responseHandler struct {
-	responseChan  chan witTypes.Result[*httpTypes.Response, httpTypes.ErrorCode]
-	responseBody  *witTypes.StreamWriter[uint8]
-	trailerWriter *witTypes.FutureWriter[witTypes.Result[witTypes.Option[*httpTypes.Fields], httpTypes.ErrorCode]]
+	responseChan chan witTypes.Result[*httpTypes.Response, httpTypes.ErrorCode]
+	bodyWriter   io.WriteCloser
 
 	httpHeaders http.Header
 
@@ -37,10 +37,10 @@ func (r *responseHandler) Header() http.Header {
 
 func (r *responseHandler) Write(b []byte) (int, error) {
 	r.headerOnce.Do(r.flush)
-	if r.responseBody == nil {
+	if r.bodyWriter == nil {
 		return 0, errors.New("response body stream is nil")
 	}
-	return int(r.responseBody.WriteAll(b)), nil
+	return r.bodyWriter.Write(b)
 }
 
 func (r *responseHandler) WriteHeader(statusCode int) {
@@ -52,36 +52,25 @@ func (r *responseHandler) WriteHeader(statusCode int) {
 
 func (r *responseHandler) Close() error {
 	r.headerOnce.Do(r.flush)
-	if r.responseBody != nil {
-		r.responseBody.Drop()
+	if r.bodyWriter != nil {
+		return r.bodyWriter.Close()
 	}
-
-	maybeTrailers := witTypes.None[*httpTypes.Fields]()
-	if len(r.httpHeaders) > 0 {
-		wasiHeaders, err := MapHttpHeaders(r.httpHeaders)
-		if err != nil {
-			return err
-		}
-		maybeTrailers = witTypes.Some(wasiHeaders)
-	}
-	r.trailerWriter.Write(witTypes.Ok[witTypes.Option[*httpTypes.Fields], httpTypes.ErrorCode](maybeTrailers))
-
 	return nil
 }
 
 func (r *responseHandler) flush() {
-	wasiHeaders, err := MapHttpHeaders(r.httpHeaders)
+	wasiHeaders, err := internalhttp.MapHttpHeaders(r.httpHeaders)
 	if err != nil {
 		// TODO: handle error
 		return
 	}
 
 	bodyWriter, bodyReader := httpTypes.MakeStreamU8()
-	r.responseBody = bodyWriter
-
 	trailerWriter, trailerFuture := httpTypes.MakeFutureResultOptionFieldsErrorCode()
-	r.trailerWriter = trailerWriter
+
+	// Reset httpHeaders so callers can populate trailers between flush and Close.
 	r.httpHeaders = make(http.Header)
+	r.bodyWriter = internalhttp.NewBodyWriter(bodyWriter, trailerWriter, r.httpHeaders)
 
 	// create response
 	resp, send := httpTypes.ResponseNew(
@@ -96,21 +85,7 @@ func (r *responseHandler) flush() {
 	r.responseChan <- witTypes.Ok[*httpTypes.Response, httpTypes.ErrorCode](resp)
 }
 
-// awaitResponse returns a single
+// awaitResponse returns a single result from the response channel.
 func (r *responseHandler) awaitResponse() witTypes.Result[*httpTypes.Response, httpTypes.ErrorCode] {
 	return <-r.responseChan
-}
-
-func MapHttpHeaders(h http.Header) (*httpTypes.Fields, error) {
-	output := httpTypes.MakeFields()
-	for key, vals := range h {
-		values := make([][]uint8, len(vals))
-		for i, val := range vals {
-			values[i] = []byte(val)
-		}
-		if res := output.Set(key, values); res.IsErr() {
-			return nil, fmt.Errorf("failed to set header %s - %+v", key, res.Err())
-		}
-	}
-	return output, nil
 }
