+++ p3/net/wasihttp/response.go
@@ -3,10 +3,9 @@ package wasihttp
 import (
 	"io"
 	"net/http"
-	"sync"
 
 	httpTypes "github.com/jamesstocktonj1/componentize-sdk/p3/gen/wasi_http_types"
-	witTypes "go.bytecodealliance.org/pkg/wit/types"
+	internalhttp "github.com/jamesstocktonj1/componentize-sdk/p3/internal/wasihttp"
 )
 
 func parseHttpResponse(req *http.Request, resp *httpTypes.Response) (*http.Response, error) {
@@ -32,59 +31,5 @@ func newResponseBodyTrailer(resp *httpTypes.Response) (io.ReadCloser, http.Heade
 	stream, trailersFut := httpTypes.ResponseConsumeBody(resp, read)
 
 	trailerMap := http.Header{}
-	return &responseBody{
-		stream:      stream,
-		trailersFut: trailersFut,
-		fut:         fut,
-		trailerMap:  trailerMap,
-	}, trailerMap
-}
-
-type responseBody struct {
-	stream      *witTypes.StreamReader[uint8]
-	trailersFut *witTypes.FutureReader[witTypes.Result[witTypes.Option[*httpTypes.Fields], httpTypes.ErrorCode]]
-	fut         *witTypes.FutureWriter[witTypes.Result[witTypes.Unit, httpTypes.ErrorCode]]
-	trailerMap  http.Header
-	headerOnce  sync.Once
-}
-
-var _ io.ReadCloser = (*responseBody)(nil)
-
-func (s *responseBody) Read(p []byte) (int, error) {
-	n := int(s.stream.Read(p))
-
-	if s.stream.WriterDropped() {
-		s.headerOnce.Do(s.readTrailers)
-		if n > 0 {
-			return n, io.EOF
-		}
-		return 0, io.EOF
-	}
-
-	return n, nil
-}
-
-func (s *responseBody) readTrailers() {
-	s.fut.Write(witTypes.Ok[witTypes.Unit, httpTypes.ErrorCode](witTypes.Unit{}))
-	result := s.trailersFut.Read()
-	if result.IsOk() {
-		opt := result.Ok()
-		if opt.IsSome() {
-			fields := opt.Some()
-			for _, kv := range fields.CopyAll() {
-				s.trailerMap.Add(kv.F0, string(kv.F1))
-			}
-			fields.Drop()
-		}
-	}
-}
-
-func (s *responseBody) Close() error {
-	// If EOF was never reached, signal completion to the host without reading trailers.
-	s.headerOnce.Do(func() {
-		s.fut.Write(witTypes.Ok[witTypes.Unit, httpTypes.ErrorCode](witTypes.Unit{}))
-		s.trailersFut.Drop()
-	})
-	s.stream.Drop()
-	return nil
+	return internalhttp.NewBodyReader(stream, trailersFut, fut, trailerMap), trailerMap
 }
